ai-course-code/chapter12_agent_basics/personal_assistant_agent/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
python-dotenv>=1.0.0
